Don't waste time under -setupInteraction under -initWithFrame for unparented WKWebViews https://bugs.webkit.org/show_bug.cgi?id=190801 <rdar://problem/43674361>
Reviewed by Megan Gardner.

* UIProcess/ios/WKContentView.mm:
(-[WKContentView _commonInitializationWithProcessPool:configuration:]):
(-[WKContentView didMoveToWindow]):
Defer the first call to WKContentViewInteraction's -setupInteraction
until the view is parented. This avoids a few milliseconds of unnecessary
work for views that are never parented.

* UIProcess/ios/WKContentViewInteraction.h:
* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView setupInteraction]):
(-[WKContentView cleanupInteraction]):
Keep track of the current state of WKContentViewInteraction's gestures.
Use this to make it OK to call -setupInteraction multiple times.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@237331 268f45cc-cd09-0410-ab3c-d52691b4dbfc