[MSE][WTF][Media] Invalid MediaTime should be falsy https://bugs.webkit.org/show_bug.cgi?id=190893
Reviewed by Jer Noble.

Source/WTF:

This patch modifies the definition of MediaTime so that invalid times
are evaluated to false in the context of a boolean expression.

* wtf/MediaTime.cpp:
(WTF::MediaTime::operator! const):
(WTF::MediaTime::operator bool const):

Tools:

This patch adds additional API tests for the conversions from
MediaTime to boolean.

* TestWebKitAPI/Tests/WTF/MediaTime.cpp:
(TestWebKitAPI::TEST):

LayoutTests:

A test is added that replicates a bug fixed by this patch.

* media/media-source/media-source-timestampoffset-then-zero-expected.txt: Added.
* media/media-source/media-source-timestampoffset-then-zero.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@237450 268f45cc-cd09-0410-ab3c-d52691b4dbfc