[GTK][WPE] Fixes to the PublicSuffix implementation https://bugs.webkit.org/show_bug.cgi?id=191031
Reviewed by Michael Catanzaro.

Source/WebCore:

Covered by existing tests.

Downcase hostnames before passing it on to the underlying
libsoup API. Special case localhost and fix a mixed-up
libsoup GError checks. This fixes most of the failures.

* platform/soup/PublicSuffixSoup.cpp:
(WebCore::isPublicSuffix):
(WebCore::topPrivatelyControlledDomain):

Tools:

* TestWebKitAPI/Tests/WebCore/PublicSuffix.cpp:
(TestWebKitAPI::TEST_F): Remove a test for a domain
that is not registrable and shouldn't be tested there.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@237584 268f45cc-cd09-0410-ab3c-d52691b4dbfc