[MSE] Use tolerance when growing the coded frame group https://bugs.webkit.org/show_bug.cgi?id=190085
Reviewed by Jer Noble.

Source/WebCore:

Test: media/media-source/media-source-append-acb-tolerance.html

This patch introduces a millisecond tolerance in the range of
potential frames that should be erased frame from the track buffer
when the coded frame group is growing.

This is necessary because some files have imprecise overlapping
timestamps (especially WebM files).

This fixes a stall when seeking back and forth in YouTube with WebM
video.

A test case simulating the problem with video/mock using timestamps
similar to those of a typical 30 fps WebM video is also added.

* Modules/mediasource/SourceBuffer.cpp:
(WebCore::SourceBuffer::sourceBufferPrivateDidReceiveSample):

LayoutTests:

A test simulating unordered appends with imprecise timestamps,
overlapping <1ms (replicating a typical WebM 30fps video file) is
added.

* media/media-source/media-source-append-acb-tolerance.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@237657 268f45cc-cd09-0410-ab3c-d52691b4dbfc