[style] Exclude style check for auto generated files. https://bugs.webkit.org/show_bug.cgi?id=191118
Reviewed by Eric Carlson.

Some files checked in to the repository are generated by other tools. In this patch,
"MiniBrowserLibResource.h" is the file generated and managed by resource editor of
VisualStudio.

There's no need to check these files because it's not practical to fix the style errors
or it's hard to maintain those fixes.

* Scripts/webkitpy/style/checkers/cpp.py:
(is_generated_file):
(CppChecker.check):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@237826 268f45cc-cd09-0410-ab3c-d52691b4dbfc