Plain text drag in contenteditable is always DragOperationCopy, never DragOperationMove https://bugs.webkit.org/show_bug.cgi?id=191228 <rdar://problem/45786830>
Patch by Jonathan Hammer <jonathan@e3software.com> on 2018-11-08
Reviewed by Wenson Hsieh.

Source/WebCore:

DragController::beginDrag should not call cleanupAfterSystemDrag because
the drag is still in progress even after the call to m_client.beginDrag()
returns. This is in contrast to DragController::doSystemDrag, where the
call to cleanupAfterSystemDrag is appropriate because the drag has
concluded by the time m_client.startDrag() returns.

Test: fast/events/drag-and-drop-move-not-copy.html

* page/DragController.cpp:
(WebCore::DragController::beginDrag):

LayoutTests:

A test that checks that dragging plain text in a contenteditable element results
in the text being moved and not copied. Only applies to WK1. Test skipped on
WK2 because WK2 EventSender does not support drag events. Test also skipped on
AppleWin because drag and drop doesn't work correctly in DRT (according to the
TestExpectations file).

* fast/events/drag-and-drop-move-not-copy-expected.txt: Added.
* fast/events/drag-and-drop-move-not-copy.html: Added.
* platform/win/TestExpectations:
* platform/wk2/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@237986 268f45cc-cd09-0410-ab3c-d52691b4dbfc