Normal-flow-only flex items don't correctly respect z-index https://bugs.webkit.org/show_bug.cgi?id=191486
Reviewed by Simon Fraser.

Source/WebCore:

Test: css3/flexbox/z-index-with-normal-flow-only.html

* rendering/RenderLayer.cpp:
(WebCore::canCreateStackingContext):
r125693 did not ensure that flex items which would otherwise be
normal-flow-only would be put into the z-order tree when necessary.
Fix by respecting the same trigger we use to make layers for flex items;
namely, not having auto z-index.

LayoutTests:

* css3/flexbox/z-index-with-normal-flow-only-expected.html: Added.
* css3/flexbox/z-index-with-normal-flow-only.html: Added.
Add a test that a <canvas> with z-index 50 correctly stacks below
a <canvas> that is a flex-item with z-index 100.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@238058 268f45cc-cd09-0410-ab3c-d52691b4dbfc