Overlay with -webkit-overflow-scrolling:touch doesn't become scrollable after added text makes it taller https://bugs.webkit.org/show_bug.cgi?id=158342 rdar://problem/26652811
Reviewed by Zalan Bujtas.
Source/WebCore:

Patch partly by Frédéric Wang.

This commit fixes an issue when resizing the content of a -webkit-overflow-scrolling: touch
overflow node on iOS. Indeed, the RenderLayerBacking's scrolling layer may not be properly
created and hence the UIProcess receives a null UIScrollView pointer. This triggers an
assertion in debug mode and prevents the user from scrolling the overflow node in release
mode. This was partially fixed by the refactoring of bug 90342 but this commit addresses
the remaining issues by forcing a configuration update after layout in order to ensure that
RenderLayerBacking's scrolling layer is available. For an overflow element that is not yet
composited, trigger a post-layout update that is necessary to check if we need to make it
composited when it gains scrollable overflow.

Tests: fast/scrolling/ios/change-scrollability-on-content-resize-nested.html
       fast/scrolling/ios/change-scrollability-on-content-resize.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::updateScrollInfoAfterLayout): Force a configuration update so that
RenderLayerCompositor::updateBackingAndHierarchy will later instantiate
RenderLayerBacking::m_scrollingLayer.

LayoutTests:

Patch partly by Frédéric Wang.

Add a test to check that scrollability of composited/non-composited overflow divs is properly
updated. This test used to fail and assert before #238090. A second test with nested divs
stills exhibit the issue and is fixed by the present commit.

* fast/scrolling/ios/change-scrollability-on-content-resize-expected.txt: Added.
* fast/scrolling/ios/change-scrollability-on-content-resize-nested-expected.txt: Added.
* fast/scrolling/ios/change-scrollability-on-content-resize-nested.html: Added.
* fast/scrolling/ios/change-scrollability-on-content-resize.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@238266 268f45cc-cd09-0410-ab3c-d52691b4dbfc