REGRESSION (r238090) Composited iframes that resize from zero don't show https://bugs.webkit.org/show_bug.cgi?id=191733 rdar://problem/46107764
Reviewed by Zalan Bujtas.
Source/WebCore:

A zero-sized iframe whose contents are composited should not trigger compositing in the
parent document (see code in requiresCompositingForFrame()), but when the <iframe> element
was resized without a style change (e.g. because it's width: 100%, height: 100% and the
parent resizes), there was no code that triggered a compositing update.

Fix by having RenderLayer::updateLayerPosition() trigger an update when the size changes,
for a RenderWidget whose contents are composited.

Test: compositing/iframes/resize-from-zero-size.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::updateLayerPosition):

LayoutTests:

* compositing/iframes/resize-from-zero-size-expected.txt: Added.
* compositing/iframes/resize-from-zero-size.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@238269 268f45cc-cd09-0410-ab3c-d52691b4dbfc