ProcessSwap.NavigationWithLockedHistoryWithoutPSON API test is a flaky crash in debug https://bugs.webkit.org/show_bug.cgi?id=191795
Reviewed by Alex Christensen.

If the WebProcessPool is getting destroyed while there is a prewarmed WebContent process
that is awaiting its connnection to the network process, then we'll hit the
process->hasOneRef() assertion in the WebProcessPool destructor. This is because
NetworkProcessProxy::m_pendingConnectionReplies keeps a RefPtr to the WebProcessProxy
that is awaiting the connection. The NetworkProcessProxy does not really need to keep
the WebProcessProxy alive in this case so this patch switches to using a WeakPtr instead.

* UIProcess/Network/NetworkProcessProxy.cpp:
(WebKit::NetworkProcessProxy::getNetworkProcessConnection):
(WebKit::NetworkProcessProxy::networkProcessCrashed):
* UIProcess/Network/NetworkProcessProxy.h:
* UIProcess/WebProcessProxy.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@238335 268f45cc-cd09-0410-ab3c-d52691b4dbfc