Editable images should always return some data, even if the canvas doesn't have a size yet https://bugs.webkit.org/show_bug.cgi?id=192265 <rdar://problem/46385911>
Reviewed by Wenson Hsieh.

* UIProcess/ios/WKDrawingView.mm:
(-[WKDrawingView layoutSubviews]):
(emptyImage):
(-[WKDrawingView renderedDrawing]):
(-[WKDrawingView PNGRepresentation]):
Some clients strongly depend on there being some data in an image, even if
it's not of a usable size yet. We'll invalidate the attachment when the
canvas size changes, so it will eventually settle at a usable size (after
the first layer tree commit that includes the editable image).


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@238767 268f45cc-cd09-0410-ab3c-d52691b4dbfc