.ycm_extra_conf.py should handle XXXInlines.h https://bugs.webkit.org/show_bug.cgi?id=193055
Reviewed by Carlos Garcia Campos.

compile_commands.json database only records flags and options for cpp source files.
So, when we open headers, this database does not have any information for that.
Our .ycm_extra_conf.py alleviates this by using XXX.cpp's configuration for XXX.h.
But this mitigation does not handle XXXInlines.h well since we do not have XXXInlines.cpp.
This patch adds support for XXXInlines.h in .ycm_extra_conf.py. When XXXInlines.h is queried,
we attempt to find XXX.cpp and use the configuration if it exists.

* gtk/ycm_extra_conf.py:
(getImplementationFilename):
(FlagsForFile):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@239567 268f45cc-cd09-0410-ab3c-d52691b4dbfc