Leak of WTF::Function objects in WebCore::CryptoKeyRSA::generatePair() (64-80 bytes each) in com.apple.WebKit.WebContent running WebKit layout tests <https://webkit.org/b/193177> <rdar://problem/47072196>
Reviewed by Saam Barati.

* crypto/mac/CryptoKeyRSAMac.cpp:
(WebCore::CryptoKeyRSA::generatePair): Fix the leak by changing
raw pointers to heap-allocated __block variables to hold the
WTF::Function objects until they are consumed within the block
passed to dispatch_async().  The __block variables act like
captured variables in a C++ lambda and have the same lifetime as
the block that they are captured in.  Note that we would have to
convert the source file from C++ to Objective-C++ to use a C++
lambda functor with dispatch_async(), which creates its own
issue because the comipiler requires a copy constructor to
convert the C++ lambda to a block functor, but the copy
constructor for the C++ lambda is implicitly deleted because the
WTF::Function copy constructor is explicitly deleted.  Whew!


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@239664 268f45cc-cd09-0410-ab3c-d52691b4dbfc