[WPE] Workaround for incorrect template specialization being selected when UChar=char16_t https://bugs.webkit.org/show_bug.cgi?id=193332
Reviewed by Michael Catanzaro.

* TestWebKitAPI/Tests/WTF/StringConcatenate.cpp: When UChar is
defined as a char16_t, which changed in ICU 59, the
StringTypeAdapter<UnsignedInt, ...> overload catches casts to
unsigned short. This test is relying on the behaviour that
UChar=unsigned short, which doesn't hold across platforms and ICU
library versions. The full fix would be a special syntax for
literal characters so that these ambiguities do not arise. That
work is proposed in https://bugs.webkit.org/show_bug.cgi?id=193101.
(TestWebKitAPI::TEST):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@239920 268f45cc-cd09-0410-ab3c-d52691b4dbfc