[WebGPU] Update vertex-buffer-triangle-strip.html to actually use vertex buffer https://bugs.webkit.org/show_bug.cgi?id=193473
Reviewed by Dean Jackson and Myles Maxfield.

Source/WebCore:

Also, switch to using the inputSlot instead of the shaderLocation field, as this seems more correct.
As of now I cannot determine an analog for WebGPU's shaderLocation in Metal.

Test: Covered by vertex-buffer-triangle-strip.html. No change in behavior.

* platform/graphics/gpu/cocoa/GPURenderPipelineMetal.mm:
(WebCore::setInputStateForPipelineDescriptor): Use the inputSlot instead of shaderLocation as bufferIndex.

LayoutTests:

Refactor the vertex shader to use the vertex stage input, rather than buffer resources.

* webgpu/vertex-buffer-triangle-strip.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@240101 268f45cc-cd09-0410-ab3c-d52691b4dbfc