[GTK] Crash while filling selection data during drag and drop https://bugs.webkit.org/show_bug.cgi?id=194698
Reviewed by Michael Catanzaro.

I can't reproduce this, but it seems that m_draggingSelectionData is nullptr in fillDragData(). That can happen
when startDrag cancels a previous DND operation, because the new m_draggingSelectionData is set before the
current DND operation is cancelled, which sets it to nullptr.

* UIProcess/gtk/DragAndDropHandler.cpp:
(WebKit::DragAndDropHandler::startDrag): Finish the previous operation before setting m_draggingSelectionData.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@241659 268f45cc-cd09-0410-ab3c-d52691b4dbfc