Fix ProcessSwap.PageOverlayLayerPersistence on macOS
* TestWebKitAPI/Tests/WebKitCocoa/ProcessSwapOnNavigation.mm:
We can't find the UI-side layer on macOS because we don't use UI-side compositing.
We could force on UI-side compositing, but I feel like the test might be more
valuable exercising the normal path on each platform (though the original
bug was also UI-side specific). Though we can't validate that the overlay
layer was installed successfully, there are layout tests that do that.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@241905 268f45cc-cd09-0410-ab3c-d52691b4dbfc