[WPE][GTK] Clean up handling of WEBKIT_FORCE_COMPLEX_TEXT https://bugs.webkit.org/show_bug.cgi?id=194614
Reviewed by Carlos Garcia Campos.

This environment variable is now enabled for WPE, not just for GTK.

It is now possible to use this environment variable to enable complex text, not just to
disable it.

WebProcessPool::setAlwaysUsesComplexTextCodePath is fixed to be respected even when this
environment variable is set, although WPE and GTK will never do so.

* UIProcess/glib/WebProcessPoolGLib.cpp:
(WebKit::WebProcessPool::platformInitialize):
(WebKit::WebProcessPool::platformInitializeWebProcess):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@242042 268f45cc-cd09-0410-ab3c-d52691b4dbfc